test(line-layer): colocate antialiasing golden
